Move cra-new to cra
