--- playgrounds/cra-new/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "cra-new",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-scripts": "4.0.3",
-    "theme-ui": "^0.12.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
