--- playgrounds/cra-new/src/App.js
@@ -1,18 +1,0 @@
-/** @jsxImportSource @emotion/react */
-import React from 'react'
-
-class MyComponent$9 extends React.Component {
-  render() {
-    return <div css={{ color: 'red' }}>red</div>
-  }
-}
-
-function App() {
-  return (
-    <div className="App">
-      <MyComponent$9 />
-    </div>
-  )
-}
-
-export default App
