+++ playgrounds/cra/package.json
@@ -1,17 +1,35 @@
 {
-  "name": "example",
-  "version": "0.6.17",
+  "name": "cra-new",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
     "react": "16.14.0",
     "react-dom": "16.14.0",
-    "react-scripts": "1.1.5"
+    "react-scripts": "4.0.3",
+    "theme-ui": "^0.12.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "test:typescript": "exit 0"
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
